+++ Runtime/Scripts/BehaviourFrameworks/GOAP/Agent.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace NgoUyenNguyen.Behaviour.GOAP
+{
+    public class Agent : MonoBehaviour
+    {
+    
+    }
+}